requirements.txt
Django>=1.8,<2.0
psycopg2
celery>=4.1,<5.0
redis
djangorestframework
django-filter
markdown
# for dev mode
docutils
# Should maybe be separated:
flower